Seeder fixes, Implementation of the form part for the team report

+++ csatar-plugins/csatar/updates/builder_table_create_csatar_csatar_team_reports_scouts.php
@@ -15,7 +15,7 @@ class BuilderTableCreateCsatarCsatarTeamReportsScouts extends Migration
             $table->string('name', 255);
             $table->integer('legal_relationship_id')->index('legal_relationship_id')->unsigned();
             $table->integer('leadership_qualification_id')->index('leadership_qualification_id')->unsigned();
-            $table->integer('membership_fee')->unsigned();
+            $table->double('membership_fee')->unsigned();
             $table->primary(['team_report_id','scout_id'], 'csatar_csatar_team_report_id_scout_id_primary');
         });
     }
